requirements.txt
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.1.0
psycopg2-binary==2.9.11
pdfminer.six==20260107
requests==2.32.5
gunicorn==23.0.0